# openapi.yaml
openapi: 3.1.0
info:
  title: BetterReads API
  description: Book tracking and review API.
  contact:
    name: BetterReads
  license:
    name: Apache 2.0
    url: https://github.com/dahl-jar/betterreads/blob/main/LICENSE
  version: v1
servers:
- url: http://localhost:8080
  description: Production
security:
- bearer-jwt: []
...
paths:
...
  /api/v1/auth/verify-email:
    post:
      tags:
      - Authentication
      summary: Confirm an email address
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Email verified or replay accepted
        "400":
          description: Invalid or expired token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: Rate limited
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
  /api/v1/auth/reset-password:
    post:
      tags:
      - Authentication
      summary: Complete a password reset
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Password replaced
        "400":
          description: "Invalid, expired, or already-consumed token"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: Rate limited
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
  /api/v1/auth/resend-verification:
    post:
      tags:
      - Authentication
      summary: Resend an email-verification link
      operationId: resendVerification
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Resend email dispatched if account is unverified
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: Rate limited
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Rotate access and refresh tokens
      operationId: refresh
      parameters:
      - name: br_refresh
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: "New access token, refresh cookie rotated"
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/AuthResponse"
        "401":
          description: "Missing, expired, or already-rotated refresh token"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/auth/forgot-password:
    post:
      tags:
      - Authentication
      summary: Start a password reset
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Reset email dispatched if account exists
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: Rate limited
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
  /api/v1/me/books/{key}:
    delete:
      tags:
      - Shelf
      summary: Remove a book from the shelf
      operationId: remove_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Book removed or already absent
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    patch:
      tags:
      - Shelf
      summary: Update reading dates and note
      operationId: update
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryRequest"
        required: true
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The updated shelf entry
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ShelfEntryResponse"
        "400":
          description: Finished date before started date
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Book is not on the shelf
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/books:
    get:
      tags:
      - Catalog
      summary: Get a homepage book list
      operationId: list_1
      parameters:
      - name: list
        in: query
        required: true
        schema:
          type: string
          enum:
          - RECENTLY_ADDED
          - TOP_RATED
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 12
          maximum: 50
          minimum: 1
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookCardResponse"
      security: []
  /api/v1/books/{key}:
    get:
      tags:
      - Catalog
      summary: Get book detail by key
      operationId: getBook
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: The book
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BookDetailResponse"
        "404":
          description: No book with that key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/books/{key}/events:
    get:
      tags:
      - Catalog
      summary: Stream book detail updates by key
      operationId: streamBook
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: One book-updated event with the full detail.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/BookDetailResponse"
        "404":
          description: No book with that key
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security: []
...
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get the current authenticated user
      operationId: me
      responses:
        "503":
          description: Service Unavailable
        "200":
          description: Current user profile
          content:
            '*/*':
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UserResponse"
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    delete:
      tags:
      - Authentication
      summary: Delete the current account
      operationId: deleteMe
      responses:
        "503":
          description: Service Unavailable
        "204":
          description: Account deleted
        "401":
          description: Missing or invalid access token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
    ShelfEntryResponse:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        coverUrl:
          type: string
        status:
          type: string
          enum:
          - WANT_TO_READ
          - CURRENTLY_READING
          - FINISHED
          - DROPPED
        favorite:
          type: boolean
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        notes:
          type: string
        addedAt:
          type: string
          format: date
        averageRating:
          type: number
        myRating:
          type: integer
          format: int32
...
    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
          description: Verification token from the email link
          example: 8f2a3c4e5d6b7a8f2a3c4e5d6b7a8f2a
          maxLength: 128
          minLength: 0
      required:
      - token
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          description: Reset token from the email link
          example: 8f2a3c4e5d6b7a8f2a3c4e5d6b7a8f2a
          maxLength: 128
          minLength: 0
        newPassword:
          type: string
          description: 8-72 bytes
          example: '********'
          maxLength: 72
          minLength: 8
      required:
      - newPassword
      - token
    ResendVerificationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
          maxLength: 255
          minLength: 0
      required:
      - email
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Short-lived access JWT
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzMwMDAwMDAwfQ.signature
        user:
          $ref: "#/components/schemas/UserResponse"
    UserResponse:
      type: object
      properties:
        username:
          type: string
          example: john.doe
        email:
          type: string
          example: john.doe@example.com
        emailVerified:
          type: boolean
          example: true
        displayName:
          type: string
          example: John Doe
        avatarUrl:
          type: string
          example: https://cdn.example.com/u/john-doe.png
        bio:
          type: string
          example: Reads mostly sci-fi.
...
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
          maxLength: 255
          minLength: 0
      required:
      - email
    UpdateEntryRequest:
      type: object
      properties:
        startedAt:
          type: string
          format: date
        finishedAt:
          type: string
          format: date
        notes:
          type: string
          maxLength: 2000
          minLength: 0
...
    BookCardResponse:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        coverUrl:
          type: string
        firstPublishYear:
          type: integer
          format: int32
        averageRating:
          type: number
        ratingCount:
          type: integer
          format: int32
    BookDetailResponse:
      type: object
      properties:
        key:
          type: string
        complete:
          type: boolean
        title:
          type: string
        subtitle:
          type: string
        authors:
          type: array
          items:
            type: string
        description:
          type: string
        coverUrl:
          type: string
        firstPublishYear:
          type: integer
          format: int32
        isbn:
          type: string
        pageCount:
          type: integer
          format: int32
        language:
          type: string
        averageRating:
          type: number
        ratingCount:
          type: integer
          format: int32
        seriesName:
          type: string
        seriesPosition:
          type: integer
          format: int32
        subjects:
          type: array
          items:
            type: string
        awards:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: Access JWT from POST /api/v1/auth/login.
      scheme: bearer
      bearerFormat: JWT